refactor: Extract loading / exceptional states
