+++ apps/web/src/studio/components/workflows/index.ts
@@ -1,4 +1,4 @@
-export * from './list/LocalStudioWorkflowPanel';
+export * from './landing-page/LocalStudioWorkflowLandingPage';
 export * from './node-view/WorkflowsDetailPage';
 export * from './test-workflow/WorkflowsTestPage';
 export * from './step-editor/WorkflowsStepEditorPage';
