+++ apps/web/src/studio/components/workflows/node-view/StepNode.tsx
@@ -1,6 +1,8 @@
+import { Skeleton } from '@mantine/core';
 import { Title, type LocalizedMessage } from '@novu/novui';
 import { css, cx } from '@novu/novui/css';
 import { hstack } from '@novu/novui/patterns';
+import { token } from '@novu/novui/tokens';
 import { FC, MouseEventHandler } from 'react';
 import { truncatedFlexTextCss } from '../../../utils/shared.styles';
 
@@ -10,7 +12,9 @@ interface IStepNodeProps {
   onClick?: MouseEventHandler<HTMLButtonElement>;
 }
 
-export const StepNode: FC<IStepNodeProps> = ({ icon, title, onClick }) => {
+type StepNodeComponent = FC<IStepNodeProps> & { LoadingDisplay: FC };
+
+export const StepNode: StepNodeComponent = ({ icon, title, onClick }) => {
   return (
     <button
       className={cx(
@@ -49,3 +53,43 @@ export const StepNode: FC<IStepNodeProps> = ({ icon, title, onClick }) => {
     </button>
   );
 };
+
+StepNode.LoadingDisplay = StepNodeSkeleton;
+
+function StepNodeSkeleton() {
+  return (
+    <div
+      className={cx(
+        css({
+          w: '[240px]',
+          cursor: 'pointer',
+          '&:not(:last-of-type):after': {
+            content: '""',
+            position: 'relative',
+            borderLeft: 'dashed',
+            borderColor: 'workflow.node.connector',
+            height: 'workflow.nodes.gap',
+            display: 'block',
+            left: '[calc(50% - 1px)]',
+          },
+        })
+      )}
+    >
+      <span
+        className={hstack({
+          gap: '50',
+          p: '150',
+          shadow: 'medium',
+          bg: 'workflow.node.surface',
+          borderRadius: '150',
+          _hover: {
+            opacity: 'hover',
+          },
+        })}
+      >
+        <Skeleton height={token('lineHeights.200')} width={token('lineHeights.200')} radius={8} />
+        <Skeleton height={token('lineHeights.100')} width={'70%'} radius="md" />
+      </span>
+    </div>
+  );
+}
